+++ CHANGES.md
@@ -1,6 +1,14 @@
 Changes
 -------
 
+0.99.1
+======
+
+Bugfixes:
+  - uneven amr slices (see #16)
+  - length-0 slices (see #16)
+  - slowness of vpic reader (see #20)
+
 0.99.0
 ======